Trying to eliminate bugs: - Sensor 0.0 never initializes on first attempt - When 0.0 is connected and NUM_SENSORS is set to 8, setup() crashes

+++ robominer_drivers/src/arduino_to_olimex/Olimexino.cpp
@@ -7,7 +7,7 @@
 #include <chrono>
 
 
-void delay(uint32_t ms) {
+void delay(uint16_t ms) {
 	usleep(1000*ms);
 }
